# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreatePolicyRequest:
    properties:
      actions:
        example:
        - notify
        - delete
        items:
          type: string
        minItems: 1
        type: array
      conditions:
        additionalProperties: {}
        type: object
      description:
        example: Automatically delete EBS volumes unused for 30 days
        type: string
      name:
        example: Delete unused EBS volumes
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      provider:
        enum:
        - aws
        - azure
        - gcp
        example: aws
        type: string
      resource_types:
        example:
        - ebs_volume
        - ebs_snapshot
        items:
          type: string
        type: array
      schedule:
        example: 0 0 * * *
        type: string
    required:
    - actions
    - name
    - organization_id
    - provider
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
...
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  handler.PaginatedResponse:
    properties:
      data: {}
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
  handler.PolicyDTO:
    properties:
      actions:
        example:
        - notify
        - delete
        items:
          enum:
          - notify
          - tag
          - stop
          - delete
          type: string
        type: array
      conditions:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      description:
        example: Automatically delete EBS volumes unused for 30 days
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_enabled:
        example: true
        type: boolean
      name:
        example: Delete unused EBS volumes
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      provider:
        enum:
        - aws
        - azure
        - gcp
        example: aws
        type: string
      resource_types:
        example:
        - ebs_volume
        items:
          type: string
        type: array
      schedule:
        example: 0 0 * * *
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cloudsweep.io
    name: CloudSweep Support
    url: https://github.com/ermt0504/cloud-sweep
  description: API FinOps/GreenOps pour automatiser le nettoyage des ressources cloud
    inutilisees
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CloudSweep API
  version: "1.0"
paths:
...
  /policies:
    get:
      consumes:
      - application/json
      description: Get a paginated list of cleanup policies
      parameters:
      - description: Filter by cloud provider
        enum:
        - aws
        - azure
        - gcp
        in: query
        name: provider
        type: string
      - description: Filter by enabled status
        in: query
        name: is_enabled
        type: boolean
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.PolicyDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List policies
      tags:
      - Policies
    post:
      consumes:
      - application/json
      description: Create a new cleanup policy
      parameters:
      - description: Policy request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.PolicyDTO'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create policy
      tags:
      - Policies
  /policies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a policy
      parameters:
      - description: Policy ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete policy
      tags:
      - Policies
    get:
      consumes:
      - application/json
      description: Get a single policy by its ID
      parameters:
      - description: Policy ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.PolicyDTO'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get policy by ID
      tags:
      - Policies
    put:
      consumes:
      - application/json
      description: Update an existing policy
      parameters:
      - description: Policy ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Policy update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.PolicyDTO'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update policy
      tags:
      - Policies
  /policies/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable a policy
      parameters:
      - description: Policy ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Disable policy
      tags:
      - Policies
  /policies/{id}/enable:
    post:
      consumes:
      - application/json
      description: Enable a policy
      parameters:
      - description: Policy ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Enable policy
      tags:
      - Policies
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"